# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "simple description",
        "title": "QrMenuApi",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}/categories/{categoryID}/dishes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get info of dishes in category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Get dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dishes.getAllDishesOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}/categories/{categoryID}/dishes/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates dish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Create dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "menu params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.DishInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dishes.createDishOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}/categories/{categoryID}/dishes/{dishID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get info of dish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Get dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "dish id",
                        "name": "dishID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dishes.getDishByIDOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates dish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Update dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "dish id",
                        "name": "dishID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "dish params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.DishInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dishes.updateDishOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes dish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Delete dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "dish id",
                        "name": "dishID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dishes.deleteDishOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dishes.createDishOutput": {
            "type": "object",
            "properties": {
                "dish": {
                    "$ref": "#/definitions/entities.DishOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dishes.deleteDishOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dishes.getAllDishesOutput": {
            "type": "object",
            "properties": {
                "dishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.DishOutput"
                    }
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dishes.getDishByIDOutput": {
            "type": "object",
            "properties": {
                "dish": {
                    "$ref": "#/definitions/entities.DishOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dishes.updateDishOutput": {
            "type": "object",
            "properties": {
                "dish": {
                    "$ref": "#/definitions/entities.DishOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "entities.DishInput": {
            "type": "object",
            "required": [
                "category_id",
                "ingriedients",
                "name",
                "place",
                "price",
                "visible"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "ingriedients": {
                    "type": "string",
                    "example": "My dish ingriedients"
                },
                "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My dish name"
                },
                "place": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "number",
                    "example": 1.99
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entities.DishOutput": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_file_name": {
                    "type": "string",
                    "example": "my-dish-image.jpg"
                },
                "ingriedients": {
                    "type": "string",
                    "example": "My dish ingriedients"
                },
                "name": {
                    "type": "string",
                    "example": "My dish name"
                },
                "place": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "number",
                    "example": 1.99
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}